Renamed AddAnnoucement.js to AddAnnouncement.js and updated route import
